refactor(tests): remove move tests/ to tests/commands
