+++ test/tests/commands/file/pwd.test.ts
@@ -5,8 +5,8 @@ import { LogLevel } from "nebula-sh";
 import { describe, it, beforeEach } from "mocha";
 import { assert } from "chai";
 
-import { runCommand } from "../../utils/run";
-import { expectMessage } from "../../utils/messages";
+import { runCommand } from "../../../utils/run";
+import { expectMessage } from "../../../utils/messages";
 
 /**
  * A test case for the pwd command.
@@ -71,7 +71,7 @@ describe("pwd command", () => {
             await runCommand(commands);
 
             // Expect the message
-            expectMessage(LogLevel.Log, expectedMessage);
+            assert.isTrue(expectMessage(LogLevel.Log, expectedMessage));
         });
     });
 });
